added person credential loading screen
